Refactored a lot of the search code and started on tidying the symbol table code up
