+++ symbol/file_scope.go
@@ -1,0 +1,13 @@
+package symbol
+
+// FileScope represents the scope in a single source file, that can contain one
+// or more source classes
+type FileScope struct {
+	// The global package that the file is located in
+	Package string
+	// Every external package that is imported into the file
+	// Formatted as map[ImportedType: full.package.path]
+	Imports map[string]string
+	// The base class that is in the file
+	BaseClass *ClassScope
+}
